refactor: reduce API surface after shared dependency move
- Delete unused naturalSort/SortFunction (test-only dead code)
- Merge FolderSortFunction into folderTypes.ts
- Stop re-exporting ImageFile and createFolderEntry from feature barrels
- Keep naturalFolderSort as the only folder sort public utility
