+++ src/features/image-viewer/index.ts
@@ -1,6 +1,5 @@
 // Image Viewer Feature Exports
 
-export type { ImageFile } from '../folder-navigation/types/folderTypes';
 export { ImageDisplay } from './components/ImageDisplay';
 // Re-export component props interfaces for convenience
 export type { ImageViewerProps } from './components/ImageViewer';
@@ -13,8 +12,6 @@ export { useKeyboardHandler } from './hooks/useKeyboardHandler';
 export type { ImageContainer } from './types/ImageContainer';
 // Types
 export type { ImageHandle, ImageSource } from './types/ImageSource';
-export type { SortFunction } from './types/SortFunction';
-export { naturalSort } from './utils/sort';
 export type {
   ActionType,
   ImageDisplayProps,
