--- src/features/image-viewer/utils/sort.ts
@@ -1,7 +1,0 @@
-import type { SortFunction } from '../types/SortFunction';
-
-export const naturalSort: SortFunction = (a, b) => {
-  const [aName, bName] = [a.name, b.name];
-
-  return aName.localeCompare(bName, 'ja', { numeric: true });
-};
